Enhance trading parameters and update dependencies
- Improved `ToString` method and added `ResetForHierarchyTransition` in `TradingParameters.cs`.
- Updated package references in `FxApi.csproj`, `app.config`, and `packages.config` to newer versions.
- Added new properties and methods in `4-Token_Storage.cs` for better hierarchy management and logging.
- Enhanced client assignment handling in `6-Hierarchy_Navigator.cs`.
- Changed `InitialStake` type from `int` to `decimal` in `FxWorthMainUI.cs`.
- Updated company name and incremented assembly version in `AssemblyInfo.cs`.
- Updated testing-related package references in `FxWorth.csproj`.

+++ FxWorth/FxWorthMainUI.cs
@@ -1294,7 +1294,7 @@ namespace FxWorth
             layerConfigForm.MartingaleLevel = (int)Martingale_Level_TXT2.Value;
             layerConfigForm.HierarchyLevels = (int)Hierarchy_Levels_TXT.Value;
             layerConfigForm.MaxDrawdown = Max_Drawdown_TXT2.Value;
-            layerConfigForm.InitialStake = (int)Stake_TXT2.Value;
+            layerConfigForm.InitialStake = Stake_TXT2.Value;
 
             List<int> layerOptions = new List<int>();
             for (int i = 2; i <= Max_Depth_TXT.Value; i++)
